early work for new telemetry
