+++ src/assessments/types/requirement.ts
@@ -1,10 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { IColumn } from '@fluentui/react';
+import { Assessment } from 'assessments/types/iassessment';
 import { UniquelyIdentifiableInstances } from 'background/instance-identifier-generator';
 import { HyperlinkDefinition } from 'common/types/hyperlink-definition';
 import {
     AssessmentNavState,
+    AssessmentStoreData,
     GeneratedAssessmentInstance,
     InstanceIdToInstanceDataMap,
 } from 'common/types/store-data/assessment-result-data';
@@ -56,6 +58,10 @@ export interface Requirement {
     getInstanceStatusColumns?: () => Readonly<IColumn>[];
     getDefaultMessage?: IGetMessageGenerator;
     instanceTableHeaderType?: InstanceTableHeaderType;
+    getCompletedStepExtraTelemetryData?: (
+        assessment: Assessment,
+        storeData: AssessmentStoreData,
+    ) => any;
 }
 
 export type VisualHelperToggleConfigDeps = {
